fix: break backend-api <-> cache-api/queue-api publish-time dependency cycle
`@checkstack/cache-api` and `@checkstack/queue-api` used `Logger` and
`Migration` from `@checkstack/backend-api` as `import type` only, yet
declared `backend-api` as a runtime dependency. Harmless in the monorepo
(`workspace:*` resolves locally), but `bun publish` freezes `workspace:*`
into a concrete pin of the other package's then-current version. Because
the edge is mutual, a registry consumer must resolve
`backend-api -> cache-api -> backend-api -> ...` backward through release
history until it hits ancient versions that shipped raw `workspace:*`
ranges and a removed `@checkstack/cache-api@0.1.0` pin, which fail to
resolve. Symptom: `bun install` errors and a missing `checkstack-dev`
binary in freshly scaffolded standalone plugins.

Move `Logger` and `Migration` to `@checkstack/common` (a dependency-free
leaf). `backend-api` re-exports both, so the ~92 existing
`import { Logger, Migration } from "@checkstack/backend-api"` call sites
are unchanged. `cache-api` and `queue-api` now depend on `common`
instead of `backend-api`, removing the cycle.
