+++ core/queue-api/package.json
@@ -8,7 +8,7 @@
   "type": "module",
   "main": "src/index.ts",
   "dependencies": {
-    "@checkstack/backend-api": "workspace:*",
+    "@checkstack/common": "workspace:*",
     "zod": "^4.0.0"
   },
   "devDependencies": {
